Rewording parts of 5.4
> and gives some security protection

Removing this as we shouldn't make statements that could be interpreted as security recommendations. Esp. Because it is against the recommendations of the "[v6ops] NSA/DOD IPv6 Security Guidance" mailing list discussion, which may provide their own document and steer regulator negotiations.

For changes to Advantages and Disadvantages see #19